CryptoAutomationTest/pyproject.toml
[project]
name = "cryptoautomationtest"
version = "0.1.0"
description = ""
authors = [
    {name = "15886815@qq.com",email = "15886815@qq.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest (>=8.4.2,<9.0.0)",
    "selenium (>=4.35.0,<5.0.0)",
    "appium-python-client (>=5.2.4,<6.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
